made density a float

+++ src/glwidget.h
@@ -57,7 +57,7 @@ private:
     GLuint m_meshProgramID;
     ShaderProgram m_hairProgram;
 
-    int m_hairDensity;
+    float m_hairDensity;
 
     Ui::MainWindow *m_ui;
     QLabel *m_fpsLabel;
